We can switch from main screen to paused and game mode and such

+++ lib/gba.h
@@ -2,7 +2,7 @@
 #define MYLIB_H
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef volatile unsigned short REG;
+typedef volatile unsigned int REG;
 typedef unsigned int u32;
 typedef u16 COLOR;
 #endif
